MBS-11609 extract KubernetesApi
extract waitForCondition
remove disabled integration tests
