+++ subprojects/common/waiter/build.gradle.kts
@@ -2,3 +2,7 @@ plugins {
     id("convention.kotlin-jvm")
     id("convention.publish-kotlin-library")
 }
+
+dependencies {
+    api(projects.common.result)
+}
