+++ subprojects/gradle/kubernetes/src/main/kotlin/com/avito/k8s/Naming.kt
@@ -1,4 +1,4 @@
-package com.avito.utils.gradle
+package com.avito.k8s
 
 /**
  * Refer to official rules: https://kubernetes.io/docs/concepts/overview/working-with-objects/names/
