+++ subprojects/gradle/kubernetes/src/main/kotlin/com/avito/utils/gradle/KubernetesCredentials.kt
@@ -3,6 +3,7 @@ package com.avito.utils.gradle
 import java.io.File
 import java.io.Serializable
 
+// todo used in gradle scripts; remove it from there
 public sealed class KubernetesCredentials : Serializable {
 
     public object Empty : KubernetesCredentials() {
