+++ subprojects/gradle/kubernetes/src/testFixtures/kotlin/com/avito/k8s/model/StubPod.kt
@@ -1,10 +1,10 @@
-package com.avito.android.runner.devices.internal.kubernetes
+package com.avito.k8s.model
 
 import com.fkorotkov.kubernetes.metadata
 import com.fkorotkov.kubernetes.status
 import io.fabric8.kubernetes.api.model.Pod
 
-internal fun KubePod.Companion.createStubInstance(
+public fun KubePod.Companion.createStubInstance(
     name: String = "stub-pod-name",
     phase: String = "Running",
     ip: String? = "stub-pod-ip"
