+++ subprojects/test-runner/device-provider/impl/build.gradle.kts
@@ -1,10 +1,7 @@
-import com.avito.android.test.applyOptionalSystemProperty
-
 plugins {
     id("convention.kotlin-jvm")
     id("convention.publish-kotlin-library")
     id("convention.test-fixtures")
-    id("convention.integration-testing")
 }
 
 dependencies {
@@ -15,18 +12,11 @@ dependencies {
     implementation(projects.gradle.process)
     implementation(projects.logger.logger)
     implementation(projects.common.result)
-
-    integTestImplementation(projects.common.truthExtensions)
+    implementation(projects.common.waiter)
 
     testImplementation(testFixtures(projects.logger.logger))
     testImplementation(libs.coroutinesTest)
 
     testFixturesApi(testFixtures(projects.common.httpClient))
-}
-
-tasks.named<Test>("integrationTest").configure {
-    applyOptionalSystemProperty("kubernetesUrl")
-    applyOptionalSystemProperty("kubernetesToken")
-    applyOptionalSystemProperty("kubernetesCaCertData")
-    applyOptionalSystemProperty("kubernetesNamespace")
+    testFixturesApi(testFixtures(projects.gradle.kubernetes))
 }
