+++ subprojects/test-runner/device-provider/impl/src/main/kotlin/com/avito/android/runner/devices/internal/kubernetes/KubernetesReservationClient.kt
@@ -6,6 +6,8 @@ import com.avito.android.runner.devices.internal.EmulatorsLogsReporter
 import com.avito.android.runner.devices.internal.RemoteDevice
 import com.avito.android.runner.devices.internal.ReservationClient
 import com.avito.android.runner.devices.model.ReservationData
+import com.avito.k8s.KubernetesApi
+import com.avito.k8s.model.KubePod
 import com.avito.logger.LoggerFactory
 import com.avito.logger.create
 import com.avito.runner.service.worker.device.DeviceCoordinate
@@ -203,9 +205,7 @@ internal class KubernetesReservationClient(
                 state.deployments.close()
                 for (deploymentName in state.deployments.toList()) {
                     launch(CoroutineName("delete-deployment-$deploymentName")) {
-                        kubernetesApi.getPods(
-                            deploymentName = deploymentName
-                        ).fold(
+                        kubernetesApi.getPods(deploymentName).fold(
                             { pods ->
                                 val runningPods = pods.filter { it.phase is KubePod.PodPhase.Running }
                                 if (runningPods.isNotEmpty()) {
@@ -245,7 +245,6 @@ internal class KubernetesReservationClient(
                                 logger.warn("Can't get pods when release", error)
                             }
                         )
-
                         kubernetesApi.deleteDeployment(deploymentName)
                     }
                 }
