+++ subprojects/test-runner/device-provider/impl/src/main/kotlin/com/avito/android/runner/devices/internal/kubernetes/KubernetesReservationClientProvider.kt
@@ -1,14 +1,14 @@
 package com.avito.android.runner.devices.internal.kubernetes
 
-import com.avito.android.runner.devices.KubernetesApiProvider
 import com.avito.android.runner.devices.internal.AndroidDebugBridgeProvider
 import com.avito.android.runner.devices.internal.EmulatorsLogsReporterProvider
+import com.avito.k8s.KubernetesApiFactory
 import com.avito.logger.LoggerFactory
 import java.io.File
 
 public class KubernetesReservationClientProvider(
     private val loggerFactory: LoggerFactory,
-    private val kubernetesApiProvider: KubernetesApiProvider,
+    private val kubernetesApiFactory: KubernetesApiFactory,
     private val reservationDeploymentFactoryProvider: ReservationDeploymentFactoryProvider,
     private val emulatorsLogsReporterProvider: EmulatorsLogsReporterProvider,
     private val androidDebugBridgeProvider: AndroidDebugBridgeProvider
@@ -19,7 +19,7 @@ public class KubernetesReservationClientProvider(
     ): KubernetesReservationClient {
         return KubernetesReservationClient(
             androidDebugBridge = androidDebugBridgeProvider.provide(),
-            kubernetesApi = kubernetesApiProvider.provide(),
+            kubernetesApi = kubernetesApiFactory.create(),
             reservationDeploymentFactory = reservationDeploymentFactoryProvider.provide(),
             loggerFactory = loggerFactory,
             emulatorsLogsReporter = emulatorsLogsReporterProvider.provide(tempLogcatDir)
