+++ subprojects/test-runner/device-provider/impl/src/main/kotlin/com/avito/android/runner/devices/internal/kubernetes/UUIDDeploymentNameGenerator.kt
@@ -1,6 +1,6 @@
 package com.avito.android.runner.devices.internal.kubernetes
 
-import com.avito.utils.gradle.toValidKubernetesName
+import com.avito.k8s.toValidKubernetesName
 import java.util.UUID
 
 internal class UUIDDeploymentNameGenerator : DeploymentNameGenerator {
